Constent naming of entities and items
Within the Store and related classes, refer to objects as items. Within
the Indigo and related classes, refer to objects as entities. This is to
make it clear these are different things. Entities are what the client
wants to store in the database. Items are more generic, and are what
is stored by the Store classes. Items can be entities, but can also be
other objects we need to store to maintain the database, e.g. index
segments, database info objects etc.