Don't store word prefixes in the index
Storing all prefixes of every word dramatically increases the storage
space needed, meaning the size limits imposed by Firestore are reached
much quicker. It also decreases performance because each time we need to
load a particular segment, we have to read and parse all prefixes too.
In reality, most of the time we only want to look up a small number of
words or prefixes.

These changes introduce the concept of a word selector, which is capable
of taking a search term and specifying which words in the index should
be looked up for it. In the case of the prefixing word selector, it
picks all words in the index that are prefixed with that search term.
Prefixes therefore don't need stored in the index themselves